BGDIINF_SB-2121: Better handling of db conncetions and small refinements
- Using g to minimise the number of db connections

- Having a fallback for topics and translations

- Smaller db connection timeout

- db_caching.py has been renamed to db.py
