+++ .env.default
@@ -1,10 +1,8 @@
 BOD_DB_NAME=bod_db_name
 BOD_DB_HOST=bod_db_host
 BOD_DB_PORT=5432
-BOD_DB_USER=user
-BOD_DB_PASSWD=secret
 
 SEARCH_SPHINX_HOST=localhost
-SEARCH_SPHINX_PORT= 9312
+SEARCH_SPHINX_PORT=9312
 
 GEODATA_STAGING=prod
